+++ cmd/testrun.go
@@ -28,7 +28,7 @@ var testrunCmd = &cobra.Command{
 			defer common.Stopwatch("Task testrun lasted").Report()
 		}
 		now := time.Now()
-		zipfile := filepath.Join(os.TempDir(), fmt.Sprintf("testrun%x.zip", common.When))
+		zipfile := filepath.Join(pathlib.TempDir(), fmt.Sprintf("testrun%x.zip", common.When))
 		defer os.Remove(zipfile)
 		common.Debug("Using temporary zip file: %v", zipfile)
 		sourceDir := filepath.Dir(robotFile)
@@ -42,7 +42,7 @@ var testrunCmd = &cobra.Command{
 			pretty.Exit(2, "Error: %v", err)
 		}
 		sentinelTime := time.Now()
-		workarea := filepath.Join(os.TempDir(), fmt.Sprintf("workarea%x", common.When))
+		workarea := filepath.Join(pathlib.TempDir(), fmt.Sprintf("workarea%x", common.When))
 		defer os.RemoveAll(workarea)
 		common.Debug("Using temporary workarea: %v", workarea)
 		err = operations.Unzip(workarea, zipfile, false, true, true)